Phase-3 File
Chatbot often powered by sophisticated language models like GPT, have gained popularity in natural language processing tasks. This model can generate human like responses based on input text and are trained on vast amounts of diverse textual data. 
To effectively train such models, a crucial step is the preprocessing of the dataset. In this Python script, we outline a systematic approach to prepare a dataset for training a
chatbot